+++ ShadcnTheme/Plugin.php
@@ -211,7 +211,7 @@ class Plugin extends Base
      */
     public function getPluginVersion(): string
     {
-        return '1.0.1';
+        return '1.0.2';
     }
 
     /**